+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Enable administrators to split facility matches into new facilities [#633](https://github.com/open-apparel-registry/open-apparel-registry/pull/633)
 - Log requests made with token authentication [#646](https://github.com/open-apparel-registry/open-apparel-registry/pull/646)
 - `./scripts/resetdb` to expedite refreshing application data during development [#672](https://github.com/open-apparel-registry/open-apparel-registry/pull/672)
+- Enable searching by OAR ID from the facility search tab [#675](https://github.com/open-apparel-registry/open-apparel-registry/pull/675)
 
 ### Changed
 - Require a token for all API endpoints [#644](https://github.com/open-apparel-registry/open-apparel-registry/pull/644)